@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.3.3 <=4.0.0"
 
 dependencies:
   flutter:
@@ -32,7 +32,7 @@ dependencies:
   provider: ^6.0.3
   dartz: ^0.10.0-nullsafety.2
   equatable: ^2.0.2
-  google_fonts: ^5.1.0
+  google_fonts: ^6.2.1
   flutter_rating_bar: ^4.0.0
   cached_network_image: ^3.0.0
   get_it: ^7.1.3
