@@ -18,7 +18,7 @@ void main() {
     mockNotifier = MockTopRatedSeriesNotifier();
   });
 
-  Widget _makeTestableWidget(Widget body) {
+  Widget makeTestableWidget(Widget body) {
     return ChangeNotifierProvider<TopRatedSeriesNotifier>.value(
       value: mockNotifier,
       child: MaterialApp(
@@ -34,7 +34,7 @@ void main() {
     final progressFinder = find.byType(CircularProgressIndicator);
     final centerFinder = find.byType(Center);
 
-    await tester.pumpWidget(_makeTestableWidget(const TopRatedSeriesPage()));
+    await tester.pumpWidget(makeTestableWidget(const TopRatedSeriesPage()));
 
     expect(centerFinder, findsOneWidget);
     expect(progressFinder, findsOneWidget);
@@ -47,7 +47,7 @@ void main() {
 
     final listViewFinder = find.byType(ListView);
 
-    await tester.pumpWidget(_makeTestableWidget(const TopRatedSeriesPage()));
+    await tester.pumpWidget(makeTestableWidget(const TopRatedSeriesPage()));
 
     expect(listViewFinder, findsOneWidget);
   });
@@ -59,7 +59,7 @@ void main() {
 
     final textFinder = find.byKey(const Key('error_message'));
 
-    await tester.pumpWidget(_makeTestableWidget(const TopRatedSeriesPage()));
+    await tester.pumpWidget(makeTestableWidget(const TopRatedSeriesPage()));
 
     expect(textFinder, findsOneWidget);
   });
